requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
chromadb>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
langdetect>=1.0.9
pyyaml>=6.0

# Export functionality
reportlab>=4.0.0
python-docx>=1.0.0
markdown>=3.5.0
Pillow>=10.0.0

# Voice (optional - can be enabled via feature flags)
# streamlit-mic-recorder>=0.0.1
# gtts>=2.4.0
# pydub>=0.25.1

# UI enhancements
streamlit-option-menu>=0.3.6
